requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-rest-swagger==2.1.2
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-openapi==1.3.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.6.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
numpy==1.22.4
numpy-financial==1.0.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.8.0
scipy==1.8.1
sympy==1.10.1
